paring down PR to include an overrides ability to each spawned mocha process options
