+++ lib/index.js
@@ -30,15 +30,15 @@ var SpawnMocha = function (opts) {
   });
   var queue = async.queue(function (task, done) {
     // Setup
-    var iterationOpts = _.merge(_.clone(opts), task.iterationOpts);
+    var overrideOpts = _.merge(_.clone(opts), task.overrideOpts);
     var bin = _.isFunction(opts.bin) ? opts.bin() : opts.bin ||
     join(__dirname, '..', 'node_modules', '.bin', 'mocha');
-    var env = _.isFunction(iterationOpts.env) ? iterationOpts.env() : iterationOpts.env || process.env;
+    var env = _.isFunction(overrideOpts.env) ? overrideOpts.env() : overrideOpts.env || process.env;
     env = _.clone(env);
 
     // Generate arguments
     var args = [];
-    _(iterationOpts.flags).each(function (val, key) {
+    _(overrideOpts.flags).each(function (val, key) {
       if (_.isFunction(val)) val = val();
       args.push((key.length > 1 ? '--' : '-') + key);
       if (_.isString(val) || _.isNumber(val)) {
@@ -48,8 +48,8 @@ var SpawnMocha = function (opts) {
 
     var stdout = newStreamBuffer();
     var stderr = newStreamBuffer();
-    var fsStream = iterationOpts.fileOutput ?
-      fs.createWriteStream(iterationOpts.fileOutput, {flags: 'a', encoding: 'utf8'}) : null;
+    var fsStream = overrideOpts.fileOutput ?
+      fs.createWriteStream(overrideOpts.fileOutput, {flags: 'a', encoding: 'utf8'}) : null;
 
     // Split xunit test report in several files if required
     if (env.JUNIT_REPORT_PATH) {
@@ -61,11 +61,11 @@ var SpawnMocha = function (opts) {
     if (opts.liveOutput) {
       child.stdout.pipe(split())
         .on('data', function (line) {
-          console.log((iterationOpts.liveOutputPrepend || '') + line);
+          console.log((overrideOpts.liveOutputPrepend || '') + line);
         });
       child.stderr.pipe(split())
         .on('data', function (line) {
-          console.error((iterationOpts.liveOutputPrepend || '') + line);
+          console.error((overrideOpts.liveOutputPrepend || '') + line);
         });
       if (fsStream) {
         child.stdout.pipe(fsStream);
@@ -112,12 +112,12 @@ var SpawnMocha = function (opts) {
   };
 
   var taskNum = 0;
-  this.add = function (files, iterationOpts) {
+  this.add = function (files, overrideOpts) {
     taskNum++;
     if (!_.isArray(files)) {
       files = [files];
     }
-    var task = {taskNum: taskNum, files: files, iterationOpts: iterationOpts || {}};
+    var task = {taskNum: taskNum, files: files, overrideOpts: overrideOpts || {}};
     queue.push(task, function (err) {
       if (err) {
         _this.emit('error', err, files);
@@ -152,30 +152,7 @@ var mochaStream = function mocha(opts) {
   return stream;
 };
 
-var mochaIteration = function mochaIteration(opts, files, cb) {
-  opts = opts || {};
-  var spawnMocha = new SpawnMocha(opts);
-  opts.iterations.forEach(function (iteration) {
-    spawnMocha.add(files, iteration);
-  });
-  var errors = [];
-  spawnMocha.on('error', function (err) {
-    console.error(err.toString());
-    errors.push(err);
-  }).on('end', function () {
-    if (errors.length > 0) {
-      console.error('ERROR SUMMARY: ');
-      _(errors).each(function (err) {
-        console.error(err);
-        console.error(err.stack);
-      }).value();
-      return cb(new Error('some tests failed'));
-    }
-    cb(null);
-  });
-};
 module.exports = {
   SpawnMocha: SpawnMocha,
-  mochaStream: mochaStream,
-  mochaIteration: mochaIteration
+  mochaStream: mochaStream
 };
